Moved util classes into their own package
